Итеративные реализации поиска числа Фибоначчи и возведения степень

+++ hw-03-alg-algorithms/src/main/java/ru/tikskit/hw03algalgorithms/fibonacci/FiboNumbers.java
@@ -7,7 +7,7 @@ public interface FiboNumbers {
 
     /**
      * Возвращает число Фибоначчи по его номеру
-     * @param numNo номер числа
+     * @param numNo номер числа начиная с 0
      */
-    int getNum(int numNo);
+    long getNum(long numNo);
 }
